+++ src/common/hash-functions.cpp
@@ -4,22 +4,22 @@
 
 namespace std
 {
-    size_t hash<pair<size_t, size_t>>::operator()(
-        const pair<size_t, size_t> key
+    uint64_t hash<pair<uint64_t, uint64_t>>::operator()(
+        const pair<uint64_t, uint64_t> key
     ) const {
-        return (get<0UL>(key) << (4UL * sizeof(size_t))) + get<1UL>(key);
+        return (get<0UL>(key) << (4UL * sizeof(uint64_t))) + get<1UL>(key);
     }
 
-    size_t hash<tuple<size_t, size_t>>::operator()(
-        const tuple<size_t, size_t> key
+    uint64_t hash<tuple<uint64_t, uint64_t>>::operator()(
+        const tuple<uint64_t, uint64_t> key
     ) const {
-        return (get<0UL>(key) << (4UL * sizeof(size_t))) + get<1UL>(key);
+        return (get<0UL>(key) << (4UL * sizeof(uint64_t))) + get<1UL>(key);
     }
 
-    size_t hash<std::vector<size_t>>::operator()(
-        const std::vector<size_t> vec
+    uint64_t hash<std::vector<uint64_t>>::operator()(
+        const std::vector<uint64_t> vec
     ) const {
-        size_t seed = vec.size();
+        uint64_t seed = vec.size();
         for (auto& i : vec) {
             seed ^= i + 0x9e3779b9 + (seed << 6UL) + (seed >> 2UL);
         }
