+++ src/modules/ac-fem-freq-d3/mesh-processing.cpp
@@ -15,12 +15,12 @@
 namespace numav {
 
 template <ElementOrder O>
-size_t SimulationAcFemFreqD3Tet<O>::Impl::_get_closest_point(
+uint64_t SimulationAcFemFreqD3Tet<O>::Impl::_get_closest_point(
     const std::array<Float,DIM> point_coords
 ) {
     Float minimum_distance_squared = std::numeric_limits<Float>::max();
-    size_t ni_closest = std::numeric_limits<size_t>::max();
-    for (size_t ni = 0UL; ni != _ni_count; ++ni) {
+    uint64_t ni_closest = std::numeric_limits<uint64_t>::max();
+    for (uint64_t ni = 0UL; ni != _ni_count; ++ni) {
         Float distance_squared = 
             power<2UL>(_ni_to_coords[ni][0UL] - point_coords[0UL]) +
             power<2UL>(_ni_to_coords[ni][1UL] - point_coords[1UL]) +
@@ -37,7 +37,7 @@ size_t SimulationAcFemFreqD3Tet<O>::Impl::_get_closest_point(
 template <ElementOrder O>
 void SimulationAcFemFreqD3Tet<O>::Impl::_load_bdf(const char* const path_to_mesh)
 {
-    constexpr size_t MAX_BDF_CHARACTERS_PER_LINE = 80UL;
+    constexpr uint64_t MAX_BDF_CHARACTERS_PER_LINE = 80UL;
     std::ifstream file(path_to_mesh);
     std::string line;
     line.reserve(MAX_BDF_CHARACTERS_PER_LINE);
@@ -59,10 +59,10 @@ void SimulationAcFemFreqD3Tet<O>::Impl::_load_bdf(const char* const path_to_mesh
 
     // second pass: parse data
     _ni_to_coords = fz::SafePtr<std::array<Float,3UL>>(_ni_count);
-    _sei_to_ni = fz::SafePtr<std::array<size_t,ENIS_COUNT<O>>>(_sei_count);
-    _vei_to_ni = fz::SafePtr<std::array<size_t,ENIV_COUNT<O>>>(_vei_count);
-    _sei_to_espg = fz::SafePtr<size_t>(_sei_count);
-    _vei_to_evpg = fz::SafePtr<size_t>(_vei_count);
+    _sei_to_ni = fz::SafePtr<std::array<uint64_t,ENIS_COUNT<O>>>(_sei_count);
+    _vei_to_ni = fz::SafePtr<std::array<uint64_t,ENIV_COUNT<O>>>(_vei_count);
+    _sei_to_espg = fz::SafePtr<uint64_t>(_sei_count);
+    _vei_to_evpg = fz::SafePtr<uint64_t>(_vei_count);
     auto it_sei_to_ni = _sei_to_ni.begin();
     auto it_vei_to_ni = _vei_to_ni.begin();
     auto it_sei_to_espg = _sei_to_espg.begin();
@@ -72,7 +72,7 @@ void SimulationAcFemFreqD3Tet<O>::Impl::_load_bdf(const char* const path_to_mesh
     while (std::getline(file, line)) {
         // all minus one are for zero base indexing conversion
         if (line.starts_with("GRID    ")) {
-            const size_t ni = parse<Float>(line.substr(8UL, 8UL)) - 1UL;
+            const uint64_t ni = parse<Float>(line.substr(8UL, 8UL)) - 1UL;
             _ni_to_coords[ni] = {
                 parse<Float>(line.substr(24UL, 8UL)),
                 parse<Float>(line.substr(32UL, 8UL)),
@@ -80,27 +80,27 @@ void SimulationAcFemFreqD3Tet<O>::Impl::_load_bdf(const char* const path_to_mesh
             };
         }
         else if (line.starts_with("CTRIA3  ")) {
-            const size_t espg = parse<size_t>(line.substr(16UL, 8UL));
+            const uint64_t espg = parse<uint64_t>(line.substr(16UL, 8UL));
             _existing_espg.insert(espg);
             *it_sei_to_espg = espg;
             ++it_sei_to_espg;
             *it_sei_to_ni = {
-                parse<size_t>(line.substr(24UL, 8UL)) - 1UL,
-                parse<size_t>(line.substr(32UL, 8UL)) - 1UL,
-                parse<size_t>(line.substr(40UL, 8UL)) - 1UL
+                parse<uint64_t>(line.substr(24UL, 8UL)) - 1UL,
+                parse<uint64_t>(line.substr(32UL, 8UL)) - 1UL,
+                parse<uint64_t>(line.substr(40UL, 8UL)) - 1UL
             };
             ++it_sei_to_ni;
         }
         else if (line.starts_with("CTETRA  ")) {
-            const size_t evpg = parse<size_t>(line.substr(16UL, 8UL));
+            const uint64_t evpg = parse<uint64_t>(line.substr(16UL, 8UL));
             _existing_evpg.insert(evpg);
             *it_vei_to_evpg = evpg;
             ++it_vei_to_evpg;
             *it_vei_to_ni = {
-                parse<size_t>(line.substr(24UL, 8UL)) - 1UL,
-                parse<size_t>(line.substr(32UL, 8UL)) - 1UL,
-                parse<size_t>(line.substr(40UL, 8UL)) - 1UL,
-                parse<size_t>(line.substr(48UL, 8UL)) - 1UL
+                parse<uint64_t>(line.substr(24UL, 8UL)) - 1UL,
+                parse<uint64_t>(line.substr(32UL, 8UL)) - 1UL,
+                parse<uint64_t>(line.substr(40UL, 8UL)) - 1UL,
+                parse<uint64_t>(line.substr(48UL, 8UL)) - 1UL
             };
             ++it_vei_to_ni;
         }
@@ -120,24 +120,24 @@ template<>
 void SimulationAcFemFreqD3Tet<ElementOrder::O2>::Impl::_generate_extra_nodes()
 {
     constexpr std::array<
-        std::array<size_t,2UL>, EXTRA_ENIV_COUNT<ElementOrder::O2>
+        std::array<uint64_t,2UL>, EXTRA_ENIV_COUNT<ElementOrder::O2>
     > VTX_PAIRS_VOL = {{
         {0UL,1UL}, {0UL,2UL}, {0UL,3UL}, {1UL,2UL}, {1UL,3UL}, {2UL,3UL}
     }};
 
     constexpr std::array<
-        std::array<size_t,2UL>, EXTRA_ENIS_COUNT<ElementOrder::O2>
+        std::array<uint64_t,2UL>, EXTRA_ENIS_COUNT<ElementOrder::O2>
     > VTX_PAIRS_SFC = {{ {0UL,1UL}, {0UL,2UL}, {1UL,2UL} }};
 
-    std::unordered_map<std::tuple<size_t,size_t>,size_t> idxs_extra_nodes;
+    std::unordered_map<std::tuple<uint64_t,uint64_t>,uint64_t> idxs_extra_nodes;
     
     // first pass: count extra nodes and save idx tuples
     fz::SafePtr<std::array<bool,EXTRA_ENIV_COUNT<ElementOrder::O2>>>
         is_extra_node(_vei_count);
-    for (size_t vei = 0UL; vei != _vei_count; ++vei) {
-        for (size_t i = 0UL; i != VTX_PAIRS_VOL.size(); ++i)
+    for (uint64_t vei = 0UL; vei != _vei_count; ++vei) {
+        for (uint64_t i = 0UL; i != VTX_PAIRS_VOL.size(); ++i)
         {
-            const std::tuple<size_t,size_t> tup = make_ascending_tuple(
+            const std::tuple<uint64_t,uint64_t> tup = make_ascending_tuple(
                 _vei_to_ni[vei][VTX_PAIRS_VOL[i][0UL]],
                 _vei_to_ni[vei][VTX_PAIRS_VOL[i][1UL]]
             );
@@ -161,12 +161,12 @@ void SimulationAcFemFreqD3Tet<ElementOrder::O2>::Impl::_generate_extra_nodes()
     temp.free();
     
     // second pass: create the extra nodes and assign to volume elements
-    for (size_t vei = 0UL; vei != _vei_count; ++vei) {
-        for (size_t i = 0UL; i != VTX_PAIRS_VOL.size(); ++i)
+    for (uint64_t vei = 0UL; vei != _vei_count; ++vei) {
+        for (uint64_t i = 0UL; i != VTX_PAIRS_VOL.size(); ++i)
         {   
             if (!is_extra_node[vei][i]) { continue; }
 
-            const std::tuple<size_t,size_t> tup = make_ascending_tuple(
+            const std::tuple<uint64_t,uint64_t> tup = make_ascending_tuple(
                 _vei_to_ni[vei][VTX_PAIRS_VOL[i][0UL]],
                 _vei_to_ni[vei][VTX_PAIRS_VOL[i][1UL]]
             );
@@ -182,18 +182,18 @@ void SimulationAcFemFreqD3Tet<ElementOrder::O2>::Impl::_generate_extra_nodes()
                 _ni_to_coords[std::get<0UL>(tup)][2UL],
                 _ni_to_coords[std::get<1UL>(tup)][2UL]
             );
-            const size_t idx_extra_node = idxs_extra_nodes.at(tup);
+            const uint64_t idx_extra_node = idxs_extra_nodes.at(tup);
             _ni_to_coords[idx_extra_node] = {x, y, z}; // add extra node
         }
     }
     is_extra_node.free();
 
     // third pass: assign nodes to surface elements
-    for (size_t sei = 0UL; sei != _sei_count; ++sei) {
-        for (size_t i = 0UL; i != VTX_PAIRS_SFC.size(); ++i)
+    for (uint64_t sei = 0UL; sei != _sei_count; ++sei) {
+        for (uint64_t i = 0UL; i != VTX_PAIRS_SFC.size(); ++i)
         {
             // create a tuple of the indices in ascending order
-            const std::tuple<size_t,size_t> tup = make_ascending_tuple(
+            const std::tuple<uint64_t,uint64_t> tup = make_ascending_tuple(
                 _sei_to_ni[sei][VTX_PAIRS_SFC[i][0UL]],
                 _sei_to_ni[sei][VTX_PAIRS_SFC[i][1UL]]
             );
